[refactor] #185 - MapUIKitView 파일 분리
