+++ MC3-BeyondThe3F/MC3-BeyondThe3F/Source/View/BucketView/BucketView.swift
@@ -48,7 +48,6 @@ struct BucketView: View {
                 }
                 .ignoresSafeArea(.keyboard)
             }
-            
             .background(Color.custom(.background))
             .sheet(isPresented: $showMusicPlayView) {
                 MusicPlayView()
